+++ karma.conf.js
@@ -1,7 +1,15 @@
-// useful links: https://mike-ward.net/2015/09/07/tips-on-setting-up-karma-testing-with-webpack/
 
-// var webpack = require('webpack');
-var webpackConfig = require('./webpack.test.config');
+// todo:
+// https://github.com/karma-runner/karma-chrome-launcher // puppeteer && ChromeHeadless
+
+const path = require('path');
+
+
+const webpackConfig = require('./webpack.test.config');
+webpackConfig.entry = {
+  specs: './specs.js'
+};
+
 
 // css modules: no need for tests, declare noop loader
 webpackConfig.module.rules[0].use = {
@@ -9,15 +17,32 @@ webpackConfig.module.rules[0].use = {
 };
 
 // js modules: no need to fix loader
-// webpackConfig.module.rules[1].use =
-
-// images modules: currently just removing their loader
-// webpackConfig.module.rules.pop();
+// webpackConfig.module.rules[1].use = {...//noop};
 
 // images modules: no need for tests, declare noop loader
 webpackConfig.module.rules[2].use = {
   loader: 'null-loader'
 };
+
+// add post loader for istanbul coverage loader
+webpackConfig.module.rules.push({
+  test: /\.js$|\.jsx$/,
+  use: {
+    loader: 'istanbul-instrumenter-loader',
+    options: {
+      esModules: true
+    }
+  },
+  include: path.resolve('src/'),
+  enforce: 'post',
+  exclude: /node_modules|\.spec\.js$/,
+});
+
+
+// empty common plugins for testing purposes (no html && css used)
+webpackConfig.plugins = [];
+
+
 
 module.exports = function(config) {
   config.set({
@@ -25,40 +50,33 @@ module.exports = function(config) {
     frameworks: [ 'mocha' ], // use the mocha test framework, seems like it's included before each starting tests
 
     files: [
-      'specs.js' //just load this file
+      'specs.js',            // just load this file
     ],
-    // plugins: ['karma-coverage-istanbul-reporter'],
 
     preprocessors: {
-      'specs.js': [ 'webpack', 'sourcemap' ] //preprocess with webpack and our sourcemap loader
+      'specs.js': [ 'webpack', 'sourcemap' ],                     // preprocess with webpack and our sourcemap loader
     },
 
-    // // optionally, configure the reporter
-    // coverageReporter: {
-    //   type : 'html',
-    //   dir : 'coverage/',
-    //   fixWebpackSourcePaths: true,
-    //   // includeAllSources: true,
-    //   instrumenterOptions: {
-    //     istanbul: { noCompact: true }
-    //   }
-    // },
+    reporters: [ /* 'dots', */ 'progress', 'coverage-istanbul' ], // report results in this formats
 
-    // https://webpack.js.org/loaders/istanbul-instrumenter-loader/
-    // https://github.com/karma-runner/karma-chrome-launcher // puppeteer && ChromeHeadless
-    // reporters: [ 'dots', 'progress', 'coverage' ], //report results in this format
-    reporters: [ 'progress' ], //report results in this format
+    coverageIstanbulReporter: {
+      dir: path.join(__dirname, 'coverage'),
+      reports: [ 'html', 'lcovonly', 'text-summary' ],
+      fixWebpackSourcePaths: true,
+      skipFilesWithNoCoverage: true,
+      verbose: true
+    },
+
     colors: true,
     loglevel: config.LOG_INFO,
     autoWatch: false,
-    // browsers: [ 'Chrome', 'PhantomJS' ], //run in Chrome
-    browsers: [ 'Chrome' ], //run in Chrome
-    singleRun: true, //just run once by default
+    browsers: [ 'Chrome' ],  // run in Chrome
+    singleRun: true,         // just run once by default
     autoWatchBatchDelay: 300,
 
     webpack: webpackConfig,
     webpackServer: {
-      noInfo: true //please don't spam the console when running in karma!
+      noInfo: true           // please don't spam the console when running in karma!
     }
   });
 };