+++ src/js/fn.js
@@ -1,4 +1,9 @@
+const div = (a, b) => {
+	return a / b;
+}
+
 module.exports = {
 	add: (a, b) => a + b,
-	mul: (a, b) => a * b
-}
+	mul: (a, b) => a * b,
+	div
+};