+++ src/js/fn.spec.js
@@ -1,9 +1,9 @@
 // import { describe, it } from 'mocha';
-// ^ is imported internally by karma & mocha modules
+// ^ is imported internally by karma (with mocha framework functions)
 
 const assert = require("assert");
 
-import { add, mul } from "./fn";
+import { add, mul, div } from "./fn";
 
 describe("Demo describe 'add'", () => {
     it("should add correctly", (done) => {
@@ -18,5 +18,11 @@ describe("Demo describe 'mul'", () => {
     });
     it("should mul correctly", () => {
         assert.strictEqual(mul(2, 4), 8);
+    });
+});
+
+describe("Demo describe 'div'", () => {
+    it("should div correctly", () => {
+        assert.strictEqual(div(2, 2), 1);
     });
 });