+++ src/js/index.js
@@ -21,11 +21,11 @@ class Comp {
 }
 new Comp;
 
-fetch('json/json.json', {})
-  .then((response) => response.json())
-  .then((response) => {
-    console.log(response);
-    console.log(JSON.stringify(response));
-  }, (error) => {
-    console.error(error);
-  });
+// fetch('json/json.json', {})
+//   .then((response) => response.json())
+//   .then((response) => {
+//     console.log(response);
+//     console.log(JSON.stringify(response));
+//   }, (error) => {
+//     console.error(error);
+//   });